build(parent): add skipChecks property
Adds a new property called skipChecks, which disables almost all checks
during build time. This is useful when running the verify lifecycle
again for integration tests, which will run the checks again
unnecessarily.

The approach taken is a bit different: we add a new property which acts
as a default value for pre-defined properties, e.g. license.skip,
checkstyle.skip, etc. That way you can use the new property disable all
checks, or use the existing properties to disable a single check.

When running tests in CI, skip the checks as we already run them during
the install phase.